extend kube-addons pyyaml workaround to fedora atomic
The script that sets up  kube-addons requires PyYAML, which
ansible installs on distros capable of package installation.
For CoreOS, this is worked-around by running python in a container.

This commit extends the same workaround to Fedora Atomic (the
workaround isn't needed for CentOS Atomic, because its image
already includes PyYAML.)

The commit changes the check for whether the workaround is required
from "is python here" to "is pyyaml importable." This is necessary
because Fedora does include python, so just checking for python
isn't sufficient to trigger the workaround.